fix(AutoCroesus): correct loot summary totals
also fixes loot output having prefix on every line

Closes #61

+++ src/main/kotlin/quoi/module/impl/dungeon/autocroesus/AutoCroesusData.kt
@@ -72,7 +72,7 @@ internal data class LootFilters(
     val limit: Int? = null,
 )
 
-internal data class LootSummaryItem(val id: String, val amount: Int, val unitValue: Double) {
+internal data class LootSummaryItem(val id: String, val amount: Long, val unitValue: Double) {
     val totalValue: Double get() = amount * unitValue
 }
